+++ src/apify/apify_storage_client/_key_value_store_collection_client.py
@@ -4,7 +4,8 @@ from typing import TYPE_CHECKING
 
 from typing_extensions import override
 
-from crawlee.base_storage_client import BaseKeyValueStoreCollectionClient, KeyValueStoreListPage, KeyValueStoreMetadata
+from crawlee.storage_clients._base import BaseKeyValueStoreCollectionClient
+from crawlee.storage_clients.models import KeyValueStoreListPage, KeyValueStoreMetadata
 
 if TYPE_CHECKING:
     from apify_client.clients import KeyValueStoreCollectionClientAsync